Rename landing page file
